// src/types/backend.d.ts
export interface SummaryData {
    totalToday: number;
    totalRange: number;
    customerFees: number;
    transactionCount: number;
}

export interface VolumeData {
    date?: string;
    month?: string;
    volume: number;
}

export interface Transaction {
    id: number;
    wallet: {
        id: number;
        userId: number;
        balance: number;
        heldBalance: number;
    };
    walletTransactionType: string;
    date: string;
    transferId: number | null;
    purpose: string;
    amount: number;
}